hotfix for SSL flag

+++ server-init/init.sh
@@ -132,41 +132,43 @@ echo "Adding user credentials to nginx config..."
 
 echo "TODO"
 
-echo "Enabling SSL in nginx test site config..."
+if [ $SSL_FLAG = 1 ]; then
+  echo "Enabling SSL in nginx test site config..."
 
-[ -f "/etc/nginx/sites-available/uos3test.com" ] && {
-  sudo rm -rfv /etc/nginx/sites-available/uos3test.com
-}
-sudo touch "/etc/nginx/sites-available/uos3test.com"
+  [ -f "/etc/nginx/sites-available/uos3test.com" ] && {
+    sudo rm -rfv /etc/nginx/sites-available/uos3test.com
+  }
+  sudo touch "/etc/nginx/sites-available/uos3test.com"
 
-srvr="server {"
-srvr="$srvr\n    listen 443 ssl;"
-srvr="$srvr\n    listen [::]:443 ssl;"
-srvr="$srvr\n    include $SNIP_CRT_POINT;"
-srvr="$srvr\n"
-srvr="$srvr\n    root /var/www/html;"
-srvr="$srvr\n    index index.html index.htm index.nginx-debian.html;"
-srvr="$srvr\n"
-srvr="$srvr\n    server_name uos3test.com www.uos3test.com;"
-srvr="$srvr\n"
-srvr="$srvr\n    location / {"
-srvr="$srvr\n            try_files \$uri \$uri/ =404;"
-srvr="$srvr\n    }"
-srvr="$srvr\n}"
-srvr="$srvr\n"
-srvr="$srvr\nserver {"
-srvr="$srvr\n    listen 80;"
-srvr="$srvr\n    listen [::]:80;"
-srvr="$srvr\n"
-srvr="$srvr\n    server_name uos3test.com www.uos3test.com;"
-srvr="$srvr\n"
-srvr="$srvr\n    return 302 https://\$server_name\$request_uri;"
-srvr="$srvr\n}\n"
-cat $srvr | sudo tee "/etc/nginx/sites-available/uos3test.com"
+  srvr="server {"
+  srvr="$srvr\n    listen 443 ssl;"
+  srvr="$srvr\n    listen [::]:443 ssl;"
+  srvr="$srvr\n    include $SNIP_CRT_POINT;"
+  srvr="$srvr\n"
+  srvr="$srvr\n    root /var/www/html;"
+  srvr="$srvr\n    index index.html index.htm index.nginx-debian.html;"
+  srvr="$srvr\n"
+  srvr="$srvr\n    server_name uos3test.com www.uos3test.com;"
+  srvr="$srvr\n"
+  srvr="$srvr\n    location / {"
+  srvr="$srvr\n            try_files \$uri \$uri/ =404;"
+  srvr="$srvr\n    }"
+  srvr="$srvr\n}"
+  srvr="$srvr\n"
+  srvr="$srvr\nserver {"
+  srvr="$srvr\n    listen 80;"
+  srvr="$srvr\n    listen [::]:80;"
+  srvr="$srvr\n"
+  srvr="$srvr\n    server_name uos3test.com www.uos3test.com;"
+  srvr="$srvr\n"
+  srvr="$srvr\n    return 302 https://\$server_name\$request_uri;"
+  srvr="$srvr\n}\n"
+  cat $srvr | sudo tee "/etc/nginx/sites-available/uos3test.com"
 
-[ ! -f "/etc/nginx/sites-enabled/uos3test.com" ] && {
-  sudo ln -sf /etc/nginx/sites-available/uos3test.com /etc/nginx/sites-enabled/uos3test.com
-}
+  [ ! -f "/etc/nginx/sites-enabled/uos3test.com" ] && {
+    sudo ln -sf /etc/nginx/sites-available/uos3test.com /etc/nginx/sites-enabled/uos3test.com
+  }
+fi
 
 echo "Completed! Have a nice day!"
 echo "To check if the new nginx configs work, run 'sudo nginx -t'."
